Refactor connector, allow to send message, update docs
